==> url_labeling/1_combine_domain_labels.ipynb <==
import time
import pandas as pd

start = time.time()
service_labels = pd.read_csv('../data/url_domain_1to1_mappings.csv')
news_labels = pd.read_csv('https://raw.githubusercontent.com/LazerLab/DomainDemo/refs/heads/main/data/existing_labels/news_local_natioanl_classification.csv')
print(f"Loaded datasets in {(time.time() - start):2f} seconds")

service_labels.columns = ['domain','label']
service_labels['domain'] = service_labels.domain.str.replace('www.','')
service_labels.sample(5)

news_labels.sample(5)

news_labels.classification.value_counts()

# start with service labels
combined_labels = service_labels.copy()
combined_labels['label_source'] = 'data_provenance_init'

# separate local / general news from northeastern
all_northeastern_domains = set(news_labels['domain'])
local_news_domains = set(news_labels[news_labels.classification=='local']['domain'])
news_domains = set(news_labels[news_labels.classification!='local']['domain'])

# update labels
combined_labels.loc[combined_labels['domain'].isin(news_domains), 'label'] = "News"
combined_labels.loc[combined_labels['domain'].isin(local_news_domains), 'label'] = "Local News"
combined_labels.loc[combined_labels['domain'].isin(all_northeastern_domains), 'label_source'] = "northeastern_domain_demo"

# missing domains (in northeastern news_labels but not in combined_labels yet)
missing_domains = news_labels[~news_labels['domain'].isin(combined_labels['domain'])]

domains_to_add = missing_domains[['domain', 'classification']].copy()
domains_to_add['label'] = domains_to_add['classification'].map(
    {'local': 'Local News'}).fillna('News')
domains_to_add['label_source'] = 'northeastern_domain_demo'
domains_to_add = domains_to_add.drop('classification', axis=1)

# Append missing domains to combined_labels
combined_labels = pd.concat([combined_labels, domains_to_add], ignore_index=True)

# Verify no duplicates
assert len(combined_labels['domain'].unique()) == len(combined_labels)

combined_labels.sample(10)

combined_labels.label.value_counts()

import numpy as np

combined_labels['set'] = 'train'

for label in combined_labels['label'].unique():
    label_indices = combined_labels[combined_labels['label'] == label].index.tolist()
    np.random.shuffle(label_indices)
    
    test_size = int(len(label_indices) * 0.10)
    val_size = int(len(label_indices) * 0.05)
    
    test_indices = label_indices[:test_size]
    val_indices = label_indices[test_size:test_size + val_size]
    
    combined_labels.loc[test_indices, 'set'] = 'test'
    combined_labels.loc[val_indices, 'set'] = 'val'

# clean and sort

combined_labels = combined_labels.dropna(subset=['label'])

combined_labels = combined_labels.sort_values(
    by=['label', 'label_source', 'set', 'domain'],
    ascending=[True, True, True, True]
).reset_index(drop=True)

combined_labels.to_csv('../data/combined_domain_labels_16k_splits.csv',index=False)

combined_labels.set.value_counts(normalize=True)

combined_labels.sample(10)

# Stats for Paper

# how many unique domains?
len(combined_labels)

# how many unique domains? (validation
len(combined_labels.domain.unique())

service_labels.label.value_counts()

# how many non-news websites?
len(service_labels[service_labels.label!='News'])
